+++ tests/data/fixtures/yii2_test_rt_index.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+    'row1' => [
+        'id' => 1,
+        'title' => 'title1',
+        'content' => 'content1',
+        'type_id' => 1,
+        'category' => [1],
+    ],
+    'row2' => [
+        'id' => 2,
+        'title' => 'title2',
+        'content' => 'content2',
+        'type_id' => 2,
+        'category' => [2],
+    ],
+];